Reorganize the Enhanced Inspector under `client/src/enhancedInspector/`
`client/src` is flat-dominant, and the Enhanced Inspector was the largest
feature still spread across it: 6 top-level modules, a webview script, a query
builder in the shared `queries/` folder, and 15 tests diluted among the
client's ~150 test files.

Apply the template established for `refactoring/` in 950fa5b — sources, webview
script, feature-owned `queries/`, and co-located `__tests__/` in one folder,
moved with `git mv` so history follows. No behavior change.

`enhancedInspector.ts` can no longer read its webview script with a
`client/src`-relative `fs.readFileSync`, so it goes through
`readWebviewScript`, and `.vscodeignore` whitelists the script at its new path
so it still ships in the `.vsix`.

`inspectRouter.ts` (the seam between the classic and enhanced inspectors),
`resources/enhancedInspector/`, and `enhancedInspectorPrintString.test.ts`
(enhanced-inspector by name, debugger-query by substance) stay where they are.
