+++ client/src/__tests__/optionalSupportOffer.test.ts
@@ -31,7 +31,7 @@ vi.mock('vscode', () => ({
   ConfigurationTarget: { Global: 1 },
 }));
 
-vi.mock('../enhancedInspectorCommand', () => ({
+vi.mock('../enhancedInspector/enhancedInspectorCommand', () => ({
   installEnhancedInspectorFeature: mocks.installEI,
 }));
 vi.mock('../refactoring/refactoringInstallCommand', () => ({
