+++ client/src/enhancedInspector/__tests__/enhancedInspectorColumns.test.ts
