+++ client/src/enhancedInspector/__tests__/enhancedInspectorErrorIsolation.test.ts
