+++ client/src/enhancedInspector/__tests__/enhancedInspectorGetViewSpecs.test.ts
