+++ client/src/enhancedInspector/__tests__/enhancedInspectorMeta.test.ts
