+++ client/src/enhancedInspector/__tests__/enhancedInspectorSelector.test.ts
