+++ client/src/enhancedInspector/__tests__/enhancedInspectorViewSpecs.test.ts
