+++ client/src/enhancedInspector/enhancedInspectorColumns.js
