+++ client/src/enhancedInspector/queries/getEnhancedInspectorViewSpecs.ts
@@ -1,5 +1,5 @@
-import { QueryExecutor } from './types';
-import { escapeString } from './util';
+import { QueryExecutor } from '../../queries/types';
+import { escapeString } from '../../queries/util';
 
 const VALID_SELECTOR =
   /^[a-zA-Z_][a-zA-Z0-9_]*:?$|^([a-zA-Z_][a-zA-Z0-9_]*:)+$|^[+\-*/<>=~&|@%?,]{1,2}$/;
